pyproject.toml
[project]
name = "demo-transaction-banking"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiokafka>=0.12.0",
    "fastapi>=0.116.1",
    "fastapi-cli>=0.0.11",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.0.0",
    "python-decouple>=3.8",
    "python-jose>=3.5.0",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.43",
    "uvicorn>=0.35.0",
    "alembic>=1.16.5",
    "bcrypt==4.1.3",
    "passlib==1.7.4",
]
